LargeTestBase is added.
